Add files via upload
Base Project that works. Needs Java EE, Maven and a server to work to work. To create the connection to the MongoDB the ports need to be the same, currently set to default Mongo port. We can define global variables like that in the web.xml file found under WebContent/WEB_INF/web.xml